# ByteArrayToHexViaLookup32Safe made faster