This PR makes `ByteArrayToHexViaLookup32Safe` a bit faster (~20-25%) by:

1. unrolling the loop with `ref` arithmetic
1. skipping redundant length checks

It potentially shows how this optimization can be done in other places

### Benchmarks

Before

|     Method |      Mean |    Error |   StdDev | Ratio | RatioSD | Allocated |
|----------- |----------:|---------:|---------:|------:|--------:|----------:|
| SafeLookup | 121.43 ns | 1.760 ns | 1.646 ns |  1.84 |    0.09 |     280 B |
|   HexMateA |  71.45 ns | 2.265 ns | 6.608 ns |  1.00 |    0.00 |     280 B |

After

|     Method |     Mean |    Error |   StdDev | Ratio | RatioSD | Allocated |
|----------- |---------:|---------:|---------:|------:|--------:|----------:|
| SafeLookup | 97.17 ns | 2.011 ns | 4.198 ns |  1.46 |    0.09 |     280 B |
|   HexMateA | 67.01 ns | 1.452 ns | 2.504 ns |  1.00 |    0.00 |     280 B |